Вынесла интерфейсы в отдельную папку
Создала папку интерфейсы, в ней два файла с интефейсами для задач и юзеров
