+++ src/app/todos-list/todos-card/todos-card.component.ts
@@ -1,6 +1,6 @@
 import { NgFor } from "@angular/common";
 import { Component, EventEmitter, Input, Output } from "@angular/core";
-import { Todo } from "../todos-list.component";
+import { Todo } from "../../interfaces/todo-interface";
 
 @Component ({
   selector: 'app-todo-card',
