+++ src/app/user-list/user-list.component.ts
@@ -4,38 +4,7 @@ import { UsersApiService } from "../users-api.service";
 import { UserCardComponent } from "./user-card/user-card.component";
 import { UserService } from "../user.service";
 import { CreateUserFormComponent } from "../create-user-form/create-user-form.component";
-
-export interface User {
-  id: number;
-  name: string;
-  username?: string;
-  email: string;
-  address?: {
-    street?: string;
-    suite?: string;
-    city?: string;
-    zipcode?: string;
-    geo?: {
-      lat?: string;
-      lng?: string;
-    }
-  },
-  phone?: string;
-  website: string;
-  company: {
-    name: string;
-    catchPhrase?: string;
-    bs?: string;
-}
-}
-
-export interface CreateUser {
-  id: number;
-  name: string;
-  email: string;
-  website: string;
-  companyName: string;
-}
+import { User, CreateUser } from "../interfaces/user-interface";
 
 @Component({
   selector: 'app-user-list',
